refactor(readersettings): extract shared reader/settings composables into feature/readersettings layer
Moves settings-facing symbols out of feature/reader/ into a new
feature/readersettings/ package so that feature/settings/ no longer
imports from feature/reader/, breaking the cross-feature import cycle
(issue #734).

Files moved wholesale (package declaration updated):
- BehaviorSection, DisplaySection, FormattingSection, UnifiedSlider,
  ReaderSettingsControls, ReaderSettingsSummaries, ReaderThemePalette,
  formatting/RenderCapabilities

Symbols extracted into dedicated files:
- CadenceHeroIcon (from cadence/CadenceUi.kt)
- AutoScrollToggleIcon (from autoscroll/AutoScrollUi.kt → AutoScrollIcons.kt)
- PlaybackSpeed object (from audio/PlaybackSpeedControl.kt)

All import sites updated across feature/reader/, feature/settings/,
feature/audio/, and feature/audiobook/. Zero behavior change.
