+++ app/src/main/kotlin/com/riffle/app/feature/reader/EpubReaderScreen.kt
@@ -50,7 +50,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import com.riffle.app.feature.readersettings.TocPanel
 import com.riffle.app.feature.reader.decorations.FigureBorderDecoration
-import com.riffle.app.feature.reader.formatting.RenderCapabilities
+import com.riffle.app.feature.readersettings.formatting.RenderCapabilities
+import com.riffle.app.feature.readersettings.swatchBackdropColor
 import com.riffle.app.feature.reader.highlights.shouldShowOpenInBook
 import com.riffle.app.feature.reader.highlights.shouldShowReadaloudUi
 import com.riffle.app.feature.reader.highlights.shouldShowShareHighlights
@@ -866,7 +867,7 @@ fun EpubReaderScreen(
                             if (formattingPrefs.showAutoScroll &&
                                 (orientation == ReaderOrientation.Vertical || orientation == ReaderOrientation.Continuous)
                             ) {
-                                com.riffle.app.feature.reader.autoscroll.AutoScrollToggleIcon(
+                                com.riffle.app.feature.readersettings.AutoScrollToggleIcon(
                                     isRunning = autoScrollState is com.riffle.core.domain.autoscroll.AutoScrollState.Running,
                                     onClick = {
                                         if (autoScrollState is com.riffle.core.domain.autoscroll.AutoScrollState.Running) {
