+++ app/src/main/kotlin/com/riffle/app/feature/reader/PdfReaderScreen.kt
@@ -55,7 +55,7 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleEventObserver
 import androidx.lifecycle.compose.LocalLifecycleOwner
-import com.riffle.app.feature.reader.formatting.RenderCapabilities
+import com.riffle.app.feature.readersettings.formatting.RenderCapabilities
 import com.riffle.app.feature.reader.formatting.toPdfiumPreferences
 import com.riffle.app.feature.readersettings.TocPanel
 import com.riffle.app.ui.theme.RiffleIcons
