+++ app/src/main/kotlin/com/riffle/app/feature/settings/panels/DisplaySettingsPanel.kt
@@ -3,7 +3,7 @@ package com.riffle.app.feature.settings.panels
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.riffle.app.R
-import com.riffle.app.feature.reader.DisplaySection
+import com.riffle.app.feature.readersettings.DisplaySection
 import com.riffle.core.domain.FormattingPreferences
 
 @Composable
