+++ app/src/main/kotlin/com/riffle/app/feature/settings/sections/AppBehaviorSection.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.riffle.app.R
-import com.riffle.app.feature.reader.BehaviorSection
+import com.riffle.app.feature.readersettings.BehaviorSection
 import com.riffle.app.feature.settings.SettingsSectionHeader
 
 @Composable
