+++ app/src/main/kotlin/com/riffle/app/feature/settings/sections/ReadingSection.kt
@@ -3,10 +3,10 @@ package com.riffle.app.feature.settings.sections
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import com.riffle.app.R
-import com.riffle.app.feature.reader.localizedAutoScrollSummary
-import com.riffle.app.feature.reader.localizedCadenceSummary
-import com.riffle.app.feature.reader.localizedDisplaySummary
-import com.riffle.app.feature.reader.localizedFormattingSummary
+import com.riffle.app.feature.readersettings.localizedAutoScrollSummary
+import com.riffle.app.feature.readersettings.localizedCadenceSummary
+import com.riffle.app.feature.readersettings.localizedDisplaySummary
+import com.riffle.app.feature.readersettings.localizedFormattingSummary
 import com.riffle.app.feature.settings.SettingsDrillInRow
 import com.riffle.app.feature.settings.SettingsPanel
 import com.riffle.app.feature.settings.SettingsSectionHeader
